requirements.txt
pyyaml==6.0
requests==2.28.1
python-vlc==3.0.16120
tidalapi==0.7.0rc1